requirements.txt
PyYAML==5.1
kubernetes==9.0.0
Flask==1.0.3